pyproject.toml
[project]
name = "nanobots-ddns"
version = "0.1.0"
description = ""
authors = [
    {name = "Ben Klein",email = "robobenklein@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "bottle (>=0.13.4,<0.14.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "dnspython (>=2.8.0,<3.0.0)",
    "valkey[libvalkey] (>=6.1.1,<7.0.0)",
]

[project.scripts]
nanobots = "nanobots_ddns.__main__:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"

